+++ docs/basic_soldering.md
@@ -3,17 +3,17 @@
 Basic soldering training will cover operation of the soldering iron and how it works, how to solder some common through hole components, what tools to use when fixing soldering mistakes, and how to clean up the workspace after oneself. You have the option to select one of the following PCB projects for this training:
 
 - LED Color Organ \
-    -> <a class="button is-link" href="https://docs.google.com/document/d/1t5I3AiSKU_hCJTbjtK76pYFr-UJzMELxTmffUZQe9ck/edit?usp=sharing"> Soldering manual for reference available here (TBD) </a>
+    -> [Soldering manual for reference available here](https://github.com/Amp-Lab-at-VT/website/raw/main/docs/mde_docs/Basic%20Soldering%20Training%20Manual%20(LED%20Color%20Organ).pdf') 
 
 Upon successfully completing the training, you'll not only gain invaluable soldering skills but also get to keep the PCB board you've assembled during the training!
 
 This training will not cover intermediate soldering techniques (techniques for soldering large surface mount devices). It is highly recomended you sign up for intermediate soldering after completion of basic training.
 
-Schedule a meeting with one of our mentors for basic training using the following form:
+Schedule a meeting with one of our mentors for basic training using the following link:
 
-<a class="button is-link" href="https://forms.gle/KyfVNspm22WMLAKk8">   
+<a class="button is-link" href="/soldering/booking">   
         
-    Soldering Training Sign Up Form
+    Soldering Training Sign Up
 </a>
 
 ---
