+++ docs/soldering.md
@@ -6,20 +6,23 @@ The AMP Lab provides three levels of soldering training for students:
 
 Completion of basic or higher soldering training allows the use of soldering equipment, unsupervised, in the AMP Lab. Anyone caught using soldering equipment without the basic or higher training will be denied access to the AMP Lab for at least one semester.
 
-Note: If you have completed general training in the past then you are basic certified and can directly move on to the intermediate training. If you have completed basic training before Fall 2023 you are intermediate   certified and can move on to advance soldering training.
+Note: If you have completed general training in the past then you are basic certified and can directly move on to the intermediate training. If you have completed basic training before Fall 2023 you are intermediate certified and can move on to advance soldering training.
 
 ### Soldering Training Signup
-Use the following form to sign up for a basic and intermediate soldering training session (Please note that basic training must be completed before starting intermediate soldering training):
+Use the following link or banner on top of this page to signup for solder training session. (Please note that basic training must be completed before starting intermediate soldering training):
 
-<a class="button is-link" href="https://forms.gle/KyfVNspm22WMLAKk8">   
+<a class="button is-link" href="/soldering/booking">   
     
-    Soldering Training Sign Up Form
+    Soldering Training Sign Up
 </a>
 
 Email the following address to schedule advance soldering training (Please note that intermediate training must be completed before starting advanced soldering training.)
 
-```
-amp-lab-leadership-team-g@vt.edu
-```
+
+<a class="button is-link" href="mailto:AMPLabSolderTrainingTeam@groups.office365.vt.edu">   
+    
+    AMPLabSolderTrainingTeam@groups.office365.vt.edu
+</a>
+
 Note: 
 Instructors are student volunteers. If you haven't received a reply, please send a follow-up email.
